update tests to make them less flake
