+++ test/android/get_ndk_root_test.go
@@ -1,0 +1,17 @@
+package android_testing
+
+import (
+	"testing"
+
+	"github.com/bugsnag/bugsnag-cli/pkg/android"
+	"github.com/stretchr/testify/assert"
+)
+
+func TestGetAndroidNDKRoot(t *testing.T) {
+	t.Log("Testing getting Android NDK Root")
+	results, err := android.GetAndroidNDKRoot("/opt/homebrew/share/android-commandlinetools/ndk/24.0.8215888")
+	if err != nil {
+		t.Error(err)
+	}
+	assert.Equal(t, "/opt/homebrew/share/android-commandlinetools/ndk/24.0.8215888", results, "The paths should match")
+}
